create SelectedFilterDisplay, remove makeFilterList from Queries2.js, update Grid container/ item, remove filter display from LineGraphButton.js
