+++ src/Components/LineGraphButton.js
@@ -23,26 +23,6 @@ const LineGraphButton = props => {
     if (open === "line") {
       return (
         <>
-          <div className="graph-titles-container">
-            <div className="graph-title-diplay">
-              <h1 className="graph-title">Data Series</h1>
-              <h2 className="graph-title-small">
-                {filters[0].selectedCategory}
-              </h2>
-            </div>
-            <div className="graph-title-diplay">
-              <h1 className="graph-title">Subsample</h1>
-              <h2 className="graph-title-small">
-                {filters[1].selectedCategory}
-              </h2>
-            </div>
-            {filters[2].selectedTableColumnName && (
-              <div className="graph-title-diplay">
-                <h3 className="graph-title">Additional Filter</h3>
-                <h3 className="graph-title-small">{makeFilterList()}</h3>
-              </div>
-            )}
-          </div>
           <LineGraph filter0={filters[0]} data={data} />
         </>
       );
@@ -60,26 +40,6 @@ const LineGraphButton = props => {
     if (open === "choropleth") {
       return (
         <>
-          <div className="graph-titles-container">
-            <div className="graph-title-diplay">
-              <h1 className="graph-title">Data Series</h1>
-              <h2 className="graph-title-small">
-                {filters[0].selectedCategory}
-              </h2>
-            </div>
-            <div className="graph-title-diplay">
-              <h1 className="graph-title">Subsample</h1>
-              <h2 className="graph-title-small">
-                {filters[1].selectedCategory}
-              </h2>
-            </div>
-            {filters[2].selectedTableColumnName && (
-              <div className="graph-title-diplay">
-                <h3 className="graph-title">Additional Filter</h3>
-                <h3 className="graph-title-small">{makeFilterList()}</h3>
-              </div>
-            )}
-          </div>
           <ChoroplethParent
             gqlData={data}
             filters={filters}
@@ -119,26 +79,6 @@ const LineGraphButton = props => {
     if (open === "bar") {
       return (
         <>
-          <div className="graph-titles-container">
-            <div className="graph-title-diplay">
-              <h3 className="graph-title">Data Series:</h3>
-              <h2 className="graph-title-small">
-                {filters[0].selectedCategory} |
-              </h2>
-            </div>
-            <div className="graph-title-diplay">
-              <h3 className="graph-title">Subsample:</h3>
-              <h2 className="graph-title-small">
-                {filters[1].selectedCategory} |
-              </h2>
-            </div>
-            {filters[2].selectedTableColumnName && (
-              <div className="graph-title-diplay">
-                <h3 className="graph-title">Additional Filter:</h3>
-                <h3 className="graph-title-small">{makeFilterList()}</h3>
-              </div>
-            )}
-          </div>
           <GraphParse
             data={data}
             filters={filters}
