+++ src/Components/Queries2.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { useQuery } from "@apollo/react-hooks";
 import { gql } from "apollo-boost";
 
@@ -145,20 +145,6 @@ const GetData = props => {
     );
   }
 
-  const makeFilterList = () => {
-    console.log("makeFilterList WAS CALLED");
-    return Object.keys(filters)
-      .filter(filterId => filterId >= 2)
-      .map(filterId => {
-        return (
-          <p>
-            {filters[filterId].selectedCategory} -{" "}
-            {getSelectedOption(filters, filterId)}
-          </p>
-        );
-      });
-  };
-
   return (
     <>
       <LineGraphButton
@@ -169,7 +155,6 @@ const GetData = props => {
         open={props.open}
         setOpen={props.setOpen}
         data={data}
-        makeFilterList={makeFilterList}
       />
     </>
   );
