+++ src/Components/scss/SelectedFilterDisplay.scss
@@ -1,0 +1,4 @@
+h3 {
+  font-size: 2rem;
+  padding: 1rem;
+}
